+++ src/d-table/utils/filter.js
@@ -7,10 +7,17 @@ export const search = (
   return rows.filter((row) => {
     for (let key of Object.keys(row)) {
       if (columnHashMap[key] && columnHashMap[key].filterable) {
-        if (!row[key] || typeof row[key] !== "string") {
+        if (
+          !row[key] ||
+          (typeof row[key] !== "string" && typeof row[key] !== "number")
+        ) {
           continue;
         }
-        let hayStack = caseSensitiveSearch ? row[key] : row[key].toLowerCase();
+        const stringRow =
+          typeof row[key] === "number" ? "" + row[key] : row[key];
+        let hayStack = caseSensitiveSearch
+          ? stringRow
+          : stringRow.toLowerCase();
         let needle = caseSensitiveSearch ? search : search.toLowerCase();
         if (hayStack.includes(needle)) {
           return true;
@@ -22,8 +29,8 @@ export const search = (
 };
 
 const filterMaps = {
-  Is: (dataItemToCheck, value) => dataItemToCheck == value,
-  "Is not": (dataItemToCheck, value) => dataItemToCheck != value,
+  Is: (dataItemToCheck, value) => dataItemToCheck === value,
+  "Is not": (dataItemToCheck, value) => dataItemToCheck !== value,
   "Is empty": (dataItemToCheck) =>
     dataItemToCheck === null ||
     dataItemToCheck === "" ||
@@ -32,13 +39,32 @@ const filterMaps = {
     dataItemToCheck === "0" ||
     !dataItemToCheck,
   "Is not empty": (dataItemToCheck) => !!dataItemToCheck,
-  "Is equal to": (dataItemToCheck, value) => dataItemToCheck === value,
-  "Is not equal to": (dataItemToCheck, value) => dataItemToCheck !== value,
-  "Begins with": (dataItemToCheck, value) => dataItemToCheck.startsWith(value),
-  "Ends with": (dataItemToCheck, value) => dataItemToCheck.endsWith(value),
-  Contains: (dataItemToCheck, value) => dataItemToCheck.includes(value),
-  "Does not contain": (dataItemToCheck, value) =>
-    !dataItemToCheck.includes(value),
+  "Is equal to": (dataItemToCheck, value) => dataItemToCheck == value,
+  "Is not equal to": (dataItemToCheck, value) => dataItemToCheck != value,
+  "Begins with": (dataItemToCheck, value) => {
+    if (typeof dataItemToCheck === "number") {
+      dataItemToCheck = "" + dataItemToCheck;
+    }
+    return dataItemToCheck.startsWith(value);
+  },
+  "Ends with": (dataItemToCheck, value) => {
+    if (typeof dataItemToCheck === "number") {
+      dataItemToCheck = "" + dataItemToCheck;
+    }
+    return dataItemToCheck.endsWith(value);
+  },
+  Contains: (dataItemToCheck, value) => {
+    if (typeof dataItemToCheck === "number") {
+      dataItemToCheck = "" + dataItemToCheck;
+    }
+    return dataItemToCheck.includes(value);
+  },
+  "Does not contain": (dataItemToCheck, value) => {
+    if (typeof dataItemToCheck === "number") {
+      dataItemToCheck = "" + dataItemToCheck;
+    }
+    return !dataItemToCheck.includes(value);
+  },
 };
 
 export const filter = (filter, rows) => {