+++ src/utils/allowOnlyNumbers.js
@@ -2,7 +2,6 @@ export const allowOnlyNumbers = (e, allowPeriod = false) => {
   const isNumber = isFinite(parseFloat(e.key));
   const value = e.key;
   if (value !== ".") {
-    console.log(value);
     if (
       !isNumber &&
       value !== "Delete" &&