+++ changelog.md
@@ -1,5 +1,12 @@
+- **19th July, 2023 - Eric McWinNEr 😴**
+  - Fixed bug with loader not working for table
+  - Added loader types to support the different kinds of loaders on the table
+  - Made the feature to exclude columns from the csv export to work
+  - Added an empty data placeholder for the table; added both a prop and a slot
 - **14th July, 2023 - Eric McWinNEr 😴**
   - Added a prop to the date picker to prevent users from selecting a future date
+  - Updated the table search to be able to search for numbers
+  - Updated the filter configuration rules to allow string based filtering with numbers
 - **13th July, 2023 - Eric McWinNEr 😴**
   - Fixed a minor bug with the search feature for the table component
   - Updated the dropdown component to show all the values in the dropdown when it is initially selected/opened